requirements.txt
jupyterlab
jupyterlab-drawio
jupyterlab-git
jupyterlab-pygments
lingua-language-detector==1.3.3
matplotlib
numpy
nbdev < 2
opencv-python-headless==4.8.1.78
pythainlp==4.0.2
pandas==2.1.1
fastapi==0.104.0
uvicorn[standard]==0.17.*
python-multipart==0.0.6
torch==2.1.0
